+++ src/Commands/CommandNames.php
@@ -10,4 +10,5 @@ interface CommandNames
     public const VALIDATE_ALL = 'validate:all';
     public const VALIDATE_COVERAGE = 'validate:coverage';
     public const VALIDATE_MESSAGES = 'validate:mess';
+    public const VALIDATE_STRUCTURE = 'validate:structure';
 }